Schur complements: improve lemma on Minimization of a Convex Quadratic Function
Fixes a mistake in the proof which stated that b ∉ range P implies b ∈ null P. This is not true. The proof is updated to avoid having to consider range P.

Updates the lemma to show that the sufficient condition for minimizers was in fact the complete characterization of all miminizers, using the improved section on the pseudoinverse orthogonal projection.